Prompt_Engineering_Book/EduTube_QuizBuilder/requirements.txt
youtube-transcript-api==0.6.2
python-dotenv
streamlit
pytube
SpeechRecognition
pydub
openai==0.28
